Summary: Added remaining uncommitted before the move

+++ examples/go/production_consumer/main.go
@@ -21,17 +21,17 @@ type ProductionConsumer struct {
 	zeroCopyConsumer *consumer.ZeroCopyConsumer
 	regularConsumer  *consumer.Consumer
 	announcer        blob.Announcer
-	
+
 	// State management
 	lastVersion      uint64
 	processedRecords uint64
 	startTime        time.Time
-	
+
 	// Resource management
-	ticker           *time.Ticker
-	backoffDuration  time.Duration
+	ticker             *time.Ticker
+	backoffDuration    time.Duration
 	maxBackoffDuration time.Duration
-	
+
 	// Statistics
 	stats struct {
 		sync.RWMutex
@@ -51,29 +51,29 @@ func NewProductionConsumer(consumerID string, blobStore blob.BlobStore, announce
 		backoffDuration:    100 * time.Millisecond,
 		maxBackoffDuration: 5 * time.Second,
 	}
-	
+
 	// Create consumers once during initialization
 	pc.regularConsumer = consumer.NewConsumer(
 		consumer.WithBlobRetriever(blobStore),
 		consumer.WithAnnouncer(announcer),
 		consumer.WithSerializer(internal.NewCapnProtoSerializer()),
 	)
-	
+
 	pc.zeroCopyConsumer = consumer.NewZeroCopyConsumer(
 		consumer.WithBlobRetriever(blobStore),
 		consumer.WithAnnouncer(announcer),
 	)
-	
+
 	return pc
 }
 
 func (pc *ProductionConsumer) Start(ctx context.Context) error {
 	log.Printf("🚀 Starting production consumer: %s", pc.consumerID)
-	
+
 	// Use adaptive ticker that adjusts based on data availability
 	pc.ticker = time.NewTicker(pc.backoffDuration)
 	defer pc.ticker.Stop()
-	
+
 	for {
 		select {
 		case <-ctx.Done():
@@ -84,7 +84,7 @@ func (pc *ProductionConsumer) Start(ctx context.Context) error {
 			if err := pc.processNewData(ctx); err != nil {
 				pc.recordError()
 				log.Printf("❌ Consumer %s error: %v", pc.consumerID, err)
-				
+
 				// Implement exponential backoff on errors
 				pc.increaseBackoff()
 			} else {
@@ -101,18 +101,18 @@ func (pc *ProductionConsumer) processNewData(ctx context.Context) error {
 	if !ok {
 		return fmt.Errorf("unsupported announcer type")
 	}
-	
+
 	latestVersion := goroutineAnnouncer.Latest()
 	if latestVersion <= 0 || uint64(latestVersion) <= pc.lastVersion {
 		// No new data - this is normal, not an error
 		return nil
 	}
-	
+
 	// Update activity timestamp
 	pc.stats.Lock()
 	pc.stats.lastActivity = time.Now()
 	pc.stats.Unlock()
-	
+
 	// Try zero-copy first
 	recordCount, err := pc.tryZeroCopyRead(ctx, latestVersion)
 	if err != nil {
@@ -125,42 +125,42 @@ func (pc *ProductionConsumer) processNewData(ctx context.Context) error {
 	} else {
 		pc.recordZeroCopyRead()
 	}
-	
+
 	// Update state
 	pc.lastVersion = uint64(latestVersion)
 	pc.processedRecords += uint64(recordCount)
-	
-	log.Printf("✅ Consumer %s processed version %d (%d records)", 
+
+	log.Printf("✅ Consumer %s processed version %d (%d records)",
 		pc.consumerID, latestVersion, recordCount)
-	
+
 	return nil
 }
 
-func (pc *ProductionConsumer) tryZeroCopyRead(ctx context.Context, version int) (int, error) {
+func (pc *ProductionConsumer) tryZeroCopyRead(ctx context.Context, version int64) (int, error) {
 	err := pc.zeroCopyConsumer.TriggerRefreshTo(ctx, version)
 	if err != nil {
 		return 0, err
 	}
-	
+
 	se := pc.zeroCopyConsumer.GetStateEngine()
 	if se == nil {
 		return 0, fmt.Errorf("no state engine available")
 	}
-	
+
 	return se.TotalRecords(), nil
 }
 
-func (pc *ProductionConsumer) tryRegularRead(ctx context.Context, version int) (int, error) {
+func (pc *ProductionConsumer) tryRegularRead(ctx context.Context, version int64) (int, error) {
 	err := pc.regularConsumer.TriggerRefreshTo(ctx, version)
 	if err != nil {
 		return 0, err
 	}
-	
+
 	se := pc.regularConsumer.GetStateEngine()
 	if se == nil {
 		return 0, fmt.Errorf("no state engine available")
 	}
-	
+
 	return se.TotalRecords(), nil
 }
 
@@ -169,7 +169,7 @@ func (pc *ProductionConsumer) increaseBackoff() {
 	if pc.backoffDuration > pc.maxBackoffDuration {
 		pc.backoffDuration = pc.maxBackoffDuration
 	}
-	
+
 	// Update ticker
 	pc.ticker.Stop()
 	pc.ticker = time.NewTicker(pc.backoffDuration)
@@ -177,7 +177,7 @@ func (pc *ProductionConsumer) increaseBackoff() {
 
 func (pc *ProductionConsumer) resetBackoff() {
 	pc.backoffDuration = 100 * time.Millisecond
-	
+
 	// Update ticker
 	pc.ticker.Stop()
 	pc.ticker = time.NewTicker(pc.backoffDuration)
@@ -206,20 +206,20 @@ func (pc *ProductionConsumer) recordError() {
 func (pc *ProductionConsumer) GetStats() map[string]interface{} {
 	pc.stats.RLock()
 	defer pc.stats.RUnlock()
-	
+
 	uptime := time.Since(pc.startTime)
-	
+
 	return map[string]interface{}{
-		"consumer_id":         pc.consumerID,
-		"versions_processed":  pc.stats.versionsProcessed,
-		"records_processed":   pc.processedRecords,
-		"zero_copy_reads":     pc.stats.zeroCopyReads,
-		"fallback_reads":      pc.stats.fallbackReads,
-		"errors":              pc.stats.errors,
-		"uptime_seconds":      uptime.Seconds(),
-		"last_version":        pc.lastVersion,
-		"last_activity":       pc.stats.lastActivity,
-		"current_backoff_ms":  pc.backoffDuration.Milliseconds(),
+		"consumer_id":        pc.consumerID,
+		"versions_processed": pc.stats.versionsProcessed,
+		"records_processed":  pc.processedRecords,
+		"zero_copy_reads":    pc.stats.zeroCopyReads,
+		"fallback_reads":     pc.stats.fallbackReads,
+		"errors":             pc.stats.errors,
+		"uptime_seconds":     uptime.Seconds(),
+		"last_version":       pc.lastVersion,
+		"last_activity":      pc.stats.lastActivity,
+		"current_backoff_ms": pc.backoffDuration.Milliseconds(),
 	}
 }
 
@@ -244,28 +244,28 @@ func NewHealthChecker(consumers []*ProductionConsumer) *HealthChecker {
 func (hc *HealthChecker) CheckHealth() []string {
 	var issues []string
 	now := time.Now()
-	
+
 	for _, consumer := range hc.consumers {
 		stats := consumer.GetStats()
-		
+
 		lastActivity, ok := stats["last_activity"].(time.Time)
 		if !ok || lastActivity.IsZero() {
 			issues = append(issues, fmt.Sprintf("Consumer %s: no activity recorded", consumer.consumerID))
 			continue
 		}
-		
+
 		if now.Sub(lastActivity) > hc.threshold {
-			issues = append(issues, fmt.Sprintf("Consumer %s: inactive for %v", 
+			issues = append(issues, fmt.Sprintf("Consumer %s: inactive for %v",
 				consumer.consumerID, now.Sub(lastActivity)))
 		}
-		
+
 		errors, _ := stats["errors"].(uint64)
 		if errors > 10 { // Threshold for error count
-			issues = append(issues, fmt.Sprintf("Consumer %s: high error count %d", 
+			issues = append(issues, fmt.Sprintf("Consumer %s: high error count %d",
 				consumer.consumerID, errors))
 		}
 	}
-	
+
 	return issues
 }
 
@@ -274,11 +274,11 @@ func main() {
 	blobStore := blob.NewInMemoryBlobStore()
 	announcer := blob.NewGoroutineAnnouncer()
 	defer announcer.Close()
-	
+
 	// Create context with cancellation
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	
+
 	// Create production consumers
 	var consumers []*ProductionConsumer
 	for i := 1; i <= 3; i++ {
@@ -289,10 +289,10 @@ func main() {
 		)
 		consumers = append(consumers, consumer)
 	}
-	
+
 	// Start health checker
 	healthChecker := NewHealthChecker(consumers)
-	
+
 	// Start consumers in separate goroutines
 	var wg sync.WaitGroup
 	for _, consumer := range consumers {
@@ -304,14 +304,14 @@ func main() {
 			}
 		}(consumer)
 	}
-	
+
 	// Start health monitoring
 	wg.Add(1)
 	go func() {
 		defer wg.Done()
 		ticker := time.NewTicker(10 * time.Second)
 		defer ticker.Stop()
-		
+
 		for {
 			select {
 			case <-ctx.Done():
@@ -326,36 +326,36 @@ func main() {
 				} else {
 					log.Printf("💚 All consumers healthy")
 				}
-				
+
 				// Print consumer stats
 				for _, consumer := range consumers {
 					stats := consumer.GetStats()
-					log.Printf("📊 %s: processed=%d, zero-copy=%d, fallback=%d, errors=%d", 
+					log.Printf("📊 %s: processed=%d, zero-copy=%d, fallback=%d, errors=%d",
 						consumer.consumerID,
 						stats["versions_processed"],
-						stats["zero_copy_reads"], 
+						stats["zero_copy_reads"],
 						stats["fallback_reads"],
 						stats["errors"])
 				}
 			}
 		}
 	}()
-	
+
 	// Handle graceful shutdown
 	sigChan := make(chan os.Signal, 1)
 	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
-	
+
 	log.Printf("🚀 Production consumers started. Press Ctrl+C to stop.")
-	
+
 	// Wait for shutdown signal
 	<-sigChan
 	log.Printf("📝 Shutdown signal received, stopping consumers...")
-	
+
 	// Cancel context to signal shutdown
 	cancel()
-	
+
 	// Wait for all goroutines to finish
 	wg.Wait()
-	
+
 	log.Printf("✅ All consumers stopped gracefully")
 }
